+++ docs/rules/__tests__/prefer-button.md.spec.ts
@@ -1,0 +1,23 @@
+import HtmlValidate from "../../../src/htmlvalidate";
+
+const markup: { [key: string]: string } = {};
+markup["incorrect"] = `<input type="button">
+<input type="submit">
+<input type="reset">
+<input type="image">`;
+markup["correct"] = `<button type="button"></button>
+<button type="submit"></button>
+<button type="reset"></button>`;
+
+describe("docs/rules/prefer-button.md", () => {
+	it("inline validation: incorrect", () => {
+		const htmlvalidate = new HtmlValidate({"rules":{"prefer-button":"error"}});
+		const report = htmlvalidate.validateString(markup["incorrect"]);
+		expect(report.results).toMatchSnapshot();
+	});
+	it("inline validation: correct", () => {
+		const htmlvalidate = new HtmlValidate({"rules":{"prefer-button":"error"}});
+		const report = htmlvalidate.validateString(markup["correct"]);
+		expect(report.results).toMatchSnapshot();
+	});
+});